refactor(auth): Update import statement for useAuth hook
